Move model selector below composer
